修改注释模板 调整代码规范

+++ cdqt-night-core/src/main/java/org/cdqt/night/core/session/Token.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 /**
  * Token操作类
  * 
- * @author LiuGangQiang Create in 2020/01/21
+ * @author LiuGangQiang Create in 2020/03/01
  */
 public class Token implements Serializable {
 	private static final long serialVersionUID = 1L;
@@ -15,21 +15,21 @@ public class Token implements Serializable {
 	/**
 	 * 私有化构造器
 	 *
-	 * @author LiuGangQiang Create in 2020/01/21
+	 * @author LiuGangQiang Create in 2020/03/01
 	 */
 	private Token() {
 	}
 
 	/**
-	 * Copyright © 2019 ChengDu RongXing Technology Co.Ltd All Rights Reserved.
-	 * 
-	 * @author LiuGangQiang Create in 2020/01/21
+	 * 内部工具类
+	 *
+	 * @author LiuGangQiang Create in 2020/03/01
 	 */
 	private static class SingleTokenHolder {
 		/**
 		 * {@link Token} 静态常量实例
 		 *
-		 * @author LiuGangQiang Create in 2020/01/21
+		 * @author LiuGangQiang Create in 2020/03/01
 		 */
 		private final static Token INSTANCE = new Token();
 	}
@@ -37,7 +37,7 @@ public class Token implements Serializable {
 	/**
 	 * 懒加载获取实例
 	 *
-	 * @author LiuGangQiang Create in 2020/01/21
+	 * @author LiuGangQiang Create in 2020/03/01
 	 * @return {@link SingleTokenHolder#INSTANCE} 实例
 	 */
 	public static Token getInstance() {
@@ -45,9 +45,9 @@ public class Token implements Serializable {
 	}
 
 	/**
-	 * 创建Token
+	 * 通过String创建Token
 	 *
-	 * @author LiuGangQiang Create in 2020/01/21
+	 * @author LiuGangQiang Create in 2020/03/01
 	 * @param terminal 平台标识枚举
 	 * @param identify 唯一区别串（这里传记录主键或唯一标识记录的数据即可）
 	 * @return 组合并经过Base64加密的唯一Token
@@ -61,14 +61,14 @@ public class Token implements Serializable {
 	};
 
 	/**
-	 * 创建Token
+	 * 通过Long创建Token
 	 *
-	 * @author LiuGangQiang Create in 2020/01/21
+	 * @author LiuGangQiang Create in 2020/03/01
 	 * @param terminal 平台标识枚举
 	 * @param identify 唯一区别串（这里传记录主键或唯一标识记录的数据即可）
 	 * @return 组合并经过Base64加密的唯一Token
 	 */
-	public String createToken(TerminalEnum terminal, int identify) {
+	public String createToken(TerminalEnum terminal, Long identify) {
 		StringBuffer token = new StringBuffer();
 		token.append(terminal.getFlag()).append(":");
 		token.append(UUID.randomUUID()).append(":");
@@ -79,9 +79,9 @@ public class Token implements Serializable {
 	/**
 	 * Token解码
 	 *
-	 * @author LiuGangQiang Create in 2020/01/21
-	 * @param token Token
-	 * @return token解码后的字符串
+	 * @author LiuGangQiang Create in 2020/03/01
+	 * @param token 加密前的字符串
+	 * @return 解码后的字符串
 	 */
 	public String decodeToekn(String token) {
 		String base64Pattern = "^([A-Za-z0-9+/]{4})*([A-Za-z0-9+/]{4}|[A-Za-z0-9+/]{3}=|[A-Za-z0-9+/]{2}==)$";
